// libs/westfield-rise-contentful-client/@types/generated/contentful.d.ts
export interface ICaseStudyFields {
  /** Internal Name */
  internalName: string;

  /** Card Image */
  image?: Asset | undefined;

  /** Page Key Information */
  pageKeyInformation?: IPageKeyInformation | undefined;

  /** Statistics List */
  statisticsList?: IStatisticsList | undefined;

  /** Body */
  body?: IBodyCopy | undefined;

  /** Quote */
  quote?: IQuote | undefined;

  /** Image Carousel */
  imageCarousel?: IImageCarousel | undefined;

  /** Accordion */
  accordion?: IAccordion | undefined;

  /** PDF */
  pdf?: Asset | undefined;
}

/** Can contain: Page Key Information, Statistic List, Body Copy, Quote, Image Carousel, Accordion */

export interface ICaseStudy extends Entry<ICaseStudyFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'caseStudy';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IFilterableSortableListFields {
  /** Internal Name */
  internalName: string;

  /** List Item Type */
  listItemType?: ICaseStudy | IPageKeyInformation | undefined;
}
// ...
export interface IPageKeyInformationFields {
  /** Internal Name */
  internalName: string;

  /** Brand Name */
  brandName?: string | undefined;

  /** Title */
  title: string;

  /** Text */
  text: string;

  /** Info List */
  infoList?: IKeyFacts[] | undefined;

  /** Row Variation */
  rowVariation?: boolean | undefined;

  /** Call To Action */
  callToAction?: ICallToAction | undefined;
}

/** Key Information for every Case study page separately. Can contain: brand name, title, text, a list of key-fact pairs, Call to Action */

export interface IPageKeyInformation extends Entry<IPageKeyInformationFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'pageKeyInformation';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}